[UPDATE BASE BEFORE MERGING(#50)][Protocol] Feat: Shannon lazy mode (#51)
This PR adds a `Lazy Mode` to `shannon` package, which fetches all the
required onchain data only when sending a relay. This is needed to avoid
issues caused by a stale cache if the block time is short (a few seconds
or less), e.g. for tests run on LocalNet.

Changes:
- Refactor the `shannon` package to separate caching functionality.
- Add a Caching fullnode implementation.
- Add a `Lazy` fullnode implementation.
- Add a `LazyMode` configuration option to `Shannon`.
- Disable caching if `LazyMode` flag is set.

---------

Signed-off-by: Arash <23505281+adshmh@users.noreply.github.com>
Co-authored-by: Daniel Olshansky <olshansky.daniel@gmail.com>