[Morse Support] All Morse EVM IDs to `service_alias` map (#103)
adding all morse EVMs to `service_alias.go`. This will help enable users
to launch on MORSE ASAP specifically with the F-chains cutover.

This is a requirement on behalf of Chainstack (and any live Morse
gateway) to cutover from Gateway Server -> Path in the immediate. This
also aligns nicely from a timing standpoint given that the set of chains
will be perfectly defined by the Foundation on 12/18 in accordance with
this list of chains:
https://docs.google.com/spreadsheets/d/1QWVGEuB2u5bkGfONoDNaltjny1jd9rJ_f7HZTjSGgQM/

---------

Co-authored-by: Daniel Olshansky <olshansky.daniel@gmail.com>